requirements.txt
django-oauth-toolkit==0.12.0
django-cors-middleware
django==1.11
django-cors-headers==2.4.0
mongoengine==0.16.3
django-role-permissions==2.2.1
pydes==2.0.1
pytlv==0.71
kavenegar==1.1.2
xlwt==1.3.0
python-barcode
pillow==5.3.0
python-dateutil==2.7.5
suds-jurko==0.6
zeep==3.2.0
mongoengine_goodjson==1.1.3
pycryptodome==3.7.2
cryptography==2.6
mysqlclient==1.3.14
# numpy==1.17.0
django-rq==2.1.0
git+https://github.com/Meysam-/python-jalali
cerberus==1.3.2
psycopg2-binary
pdfkit==0.6.1
django-redis==4.11.0
rq-scheduler==0.9.1
ua_parser==0.10.0
openpyxl==3.0.3
firebase-admin==4.4.0
celery==4.4.7